backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pydantic==2.10.2
python-multipart==0.0.9
